Revert "NSUInteger changed to int"
This reverts commit 64ac557822d1712afd4df03a23b689a53999ac0e.

+++ CurioSDK.Core/CurioSDK/CurioAction.h
@@ -44,7 +44,7 @@ typedef NS_ENUM(NSUInteger, CActionType) {
  
  */
 - (id) init:(NSString *) aId
-       type:(int) type
+       type:(NSUInteger) type
       stamp:(NSString *) stamp
       title:(NSString *) title
        path:(NSString *) path
